+++ test/ecto_cqs/loader_test.exs
@@ -29,26 +29,48 @@ defmodule EctoCQS.LoaderTest do
 
   test "all_ordered_by/1 returns all users ordered using given expression" do
     now = now()
+
     user_1 = insert!(:user, inserted_at: DateTime.add(now, 1, :second))
     user_2 = insert!(:user, inserted_at: DateTime.add(now, 2, :second))
 
     assert Loader.all_ordered_by(desc: :inserted_at) == [user_2, user_1]
   end
 
-  test "first/0 returns first created client" do
+  test "first/0 returns first created user" do
     now = now()
+
     user_1 = insert!(:user, inserted_at: DateTime.add(now, 1, :second))
     _user_2 = insert!(:user, inserted_at: DateTime.add(now, 2, :second))
 
     assert Loader.first() == user_1
   end
 
-  test "last/0 returns last created client" do
+  test "first/1 returns first N created users" do
+    now = now()
+
+    user_1 = insert!(:user, inserted_at: DateTime.add(now, 1, :second))
+    user_2 = insert!(:user, inserted_at: DateTime.add(now, 2, :second))
+    _user_3 = insert!(:user, inserted_at: DateTime.add(now, 3, :second))
+
+    assert Loader.first(2) == [user_1, user_2]
+  end
+
+  test "last/0 returns last created user" do
     now = now()
     _user_1 = insert!(:user, inserted_at: DateTime.add(now, 1, :second))
     user_2 = insert!(:user, inserted_at: DateTime.add(now, 2, :second))
 
     assert Loader.last() == user_2
+  end
+
+  test "last/1 returns last N created users" do
+    now = now()
+
+    _user_1 = insert!(:user, inserted_at: DateTime.add(now, 1, :second))
+    user_2 = insert!(:user, inserted_at: DateTime.add(now, 2, :second))
+    user_3 = insert!(:user, inserted_at: DateTime.add(now, 3, :second))
+
+    assert Loader.last(2) == [user_3, user_2]
   end
 
   test "count/0 returns count of all users" do