+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## v0.1.9
+
+- add `EctoCQS.Loader.first/1` and `EctoCQS.Loader.last/1`
+
 ## v0.1.8
 
 - add `EctoCQS.Loader.count_by/1`
 
 ## v0.1.5
 
-- remove `EctoCQS.Query.order_by/2` in favour of `Ecto.Query.where/3`
+- remove `EctoCQS.Query.by/2` in favour of `Ecto.Query.where/3`
 
 ## v0.1.4